Fragmented the code and its now organized
